pyproject.toml
[project]
name = "youtube-transcribe"
version = "0.1.0"
description = ""
authors = [
    {name = "Gasser",email = "gaserkhaled@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytube (>=15.0.0,<16.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "google-cloud-speech (>=2.30.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "google-cloud-storage (>=2.19.0,<3.0.0)",
    "pytubefix (>=8.12.0,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
